Add option to ignore existing files
This commit adds support to overwrite / no-overwrite file during
the conversion process. So, by default if the desire output file
is already exists in the output directory then then it'll ignore
that writing process. But, if we want to then we can force the
rewrite using the `--overwrite` option with those command.

Related: 29

+++ lib/relaton/cli/base_convertor.rb
@@ -11,6 +11,7 @@ module Relaton
         @options = options
         @outdir = options.fetch(:outdir, nil)
         @writable = options.fetch(:write, true)
+        @overwrite = options.fetch(:overwrite, true)
 
         install_dependencies(options[:require] || [])
       end
@@ -41,7 +42,7 @@ module Relaton
 
       private
 
-      attr_reader :file, :outdir, :options, :writable
+      attr_reader :file, :outdir, :options, :writable, :overwrite
 
       def default_ext
         raise "Override this method"
@@ -71,7 +72,10 @@ module Relaton
 
       def write_to_a_file(content, outfile = nil)
         outfile ||= Pathname.new(file).sub_ext(extension).to_s
-        File.open(outfile, "w:utf-8") { |file| file.write(content) }
+
+        if !File.exists?(outfile) || overwrite
+          File.open(outfile, "w:utf-8") { |file| file.write(content) }
+        end
       end
 
       def write_to_file_collection(content, format)
